[ENH] Refactor Freesurfer pipelines architecture (#1120)
* refactor freesurfer pipelines

* add forgotten cli

* use absolute imports in task definitions

* allow string or Path for get_subject_id

* fix folder creation in init_input_node
