+++ clinica/pipelines/__init__.py
@@ -10,8 +10,7 @@ from . import (
     statistics_surface,
     statistics_volume,
     statistics_volume_correction,
-    t1_freesurfer,
-    t1_freesurfer_longitudinal,
+    anatomical,
     t1_linear,
     t1_volume,
     t1_volume_create_dartel,
