+++ clinica/pipelines/anatomical/freesurfer/__init__.py
@@ -1,0 +1,1 @@
+from . import atlas, longitudinal, t1
