+++ clinica/pipelines/anatomical/freesurfer/atlas/__init__.py
@@ -1,0 +1,1 @@
+from . import cli
