+++ clinica/pipelines/anatomical/freesurfer/longitudinal/__init__.py
@@ -1,0 +1,1 @@
+from . import cli, correction, template
