+++ clinica/pipelines/anatomical/freesurfer/longitudinal/template/info.json
@@ -1,0 +1,14 @@
+{
+    "id": "aramislab/t1-freesurfer-longitudinal",
+    "author": ["Alexis Guyot", "Alexandre Routier"],
+    "version": "0.1",
+    "space_caps": "300M",
+    "space_wd": "300M",
+    "dependencies": [
+        {
+            "type": "software",
+            "name": "freesurfer",
+            "version": ">=6.0"
+        }
+    ]
+}
