+++ clinica/pipelines/anatomical/freesurfer/t1/__init__.py
@@ -1,0 +1,1 @@
+from . import cli
