--- clinica/pipelines/t1_freesurfer/__init__.py
@@ -1,1 +1,0 @@
-from . import t1_freesurfer_cli
