--- clinica/pipelines/t1_freesurfer_longitudinal/__init__.py
@@ -1,5 +1,0 @@
-from . import (
-    t1_freesurfer_longitudinal_cli,
-    t1_freesurfer_longitudinal_correction_cli,
-    t1_freesurfer_template_cli,
-)
